refactor: update import paths to use public package interfaces instead of private modules
